execute rewriteRun
