+++ app/src/main/java/io/github/pc_com/app/SlashCommand.java
@@ -1,0 +1,10 @@
+package io.github.pc_com.app;
+
+import net.dv8tion.jda.api.interactions.commands.build.SlashCommandData;
+
+/**
+ * @author kurages
+ */
+public interface SlashCommand {
+    SlashCommandData slashCommandData();
+}
