+++ app/src/main/java/io/github/pc_com/app/core/discord/internal/url/IURLFactory.java
@@ -1,0 +1,10 @@
+package io.github.pc_com.app.core.discord.internal.url;
+
+import java.net.URI;
+
+import io.github.pc_com.app.core.discord.DiscordURL;
+
+public interface IURLFactory {
+    boolean isSupported(URI url);
+    DiscordURL create(URI url);
+}
