+++ app/src/main/java/io/github/pc_com/app/system/SlashCommandRegister.java
@@ -1,0 +1,26 @@
+package io.github.pc_com.app.system;
+
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import net.dv8tion.jda.api.events.session.ReadyEvent;
+import net.dv8tion.jda.api.hooks.ListenerAdapter;
+import net.dv8tion.jda.api.requests.RestAction;
+import io.github.pc_com.app.SlashCommand;
+import org.springframework.stereotype.Component;
+
+/**
+ * @author kurages
+ */
+@Component
+@RequiredArgsConstructor
+public class SlashCommandRegister extends ListenerAdapter {
+    private final List<SlashCommand> commands;
+
+    @Override
+    public void onReady(ReadyEvent event) {
+        this.commands.stream()
+            .map(SlashCommand::slashCommandData)
+            .map(event.getJDA()::upsertCommand)
+            .forEach(RestAction::queue);
+    }
+}
