--- app/src/main/java/org/example/app/App.java
@@ -1,30 +1,0 @@
-package org.example.app;
-
-import lombok.RequiredArgsConstructor;
-import net.dv8tion.jda.api.JDA;
-import org.springframework.boot.ApplicationArguments;
-import org.springframework.boot.ApplicationRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
-import org.springframework.context.annotation.FullyQualifiedAnnotationBeanNameGenerator;
-
-/**
- * @author kurages
- */
-@RequiredArgsConstructor
-@ConfigurationPropertiesScan
-@SpringBootApplication(nameGenerator = FullyQualifiedAnnotationBeanNameGenerator.class)
-public class App implements ApplicationRunner {
-    public static void main(String[] args) {
-        SpringApplication.run(App.class, args);
-    }
-
-
-    private final JDA jda;
-
-    @Override
-    public void run(ApplicationArguments args) throws Exception {
-        jda.awaitShutdown();
-    }
-}
