--- app/src/main/java/org/example/app/DiscordConfig.java
@@ -1,16 +1,0 @@
-package org.example.app;
-
-import java.util.Objects;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-/**
- * @author kurages
- */
-@ConfigurationProperties(prefix = "discord")
-public record DiscordConfig (
-    String token
-) {
-    public DiscordConfig {
-        Objects.requireNonNull(token);
-    }
-}
