--- app/src/main/java/org/example/app/SlashCommand.java
@@ -1,10 +1,0 @@
-package org.example.app;
-
-import net.dv8tion.jda.api.interactions.commands.build.SlashCommandData;
-
-/**
- * @author kurages
- */
-public interface SlashCommand {
-    SlashCommandData slashCommandData();
-}
