--- app/src/main/java/org/example/app/core/discord/internal/url/DiscordGg.java
@@ -1,26 +1,0 @@
-package org.example.app.core.discord.internal.url;
-
-import java.net.URI;
-
-import org.example.app.core.discord.DiscordURL;
-
-import lombok.AccessLevel;
-import lombok.RequiredArgsConstructor;
-
-/**
- * 招待ショートリンク用の解析処理。
- */
-@RequiredArgsConstructor(access = AccessLevel.PACKAGE)
-class DiscordGg implements DiscordURL {
-    private final URI url;
-
-    @Override
-    public DiscordURLType getUrlType() {
-        return DiscordURLType.INVITE;
-    }
-
-    @Override
-    public String getTicket() {
-        return url.getPath().substring(1);
-    }
-}
