--- app/src/main/java/org/example/app/core/discord/internal/url/IURLFactory.java
@@ -1,10 +1,0 @@
-package org.example.app.core.discord.internal.url;
-
-import java.net.URI;
-
-import org.example.app.core.discord.DiscordURL;
-
-public interface IURLFactory {
-    boolean isSupported(URI url);
-    DiscordURL create(URI url);
-}
